+++ src/modules/dashboard/gameDev/gameDevHandlers/achievementGameDevHandler.ts
@@ -1,0 +1,59 @@
+import Boom from '@hapi/boom';
+import type { Lifecycle } from '@hapi/hapi';
+
+import { arrayToJSON } from '../../../../api-utils/utils';
+import { validateGameAuth } from '../../../../api-utils/validateGameAuth';
+import type { AchievementEditorData } from '../../../../models/Achievements';
+import {
+  findAllAchievementsByGameType,
+  getAchievementByCreator,
+  createOrUpdateAchievement,
+  deleteAchievementById,
+} from '../../../../models/Achievements';
+
+export const getAchievementsHandler: Lifecycle.Method = async (request, h) => {
+  await validateGameAuth(request.pre.getAuthUser.id, request.params.gameTypeId);
+
+  if (request.params.achievementId) {
+    const achievement = await getAchievementByCreator(
+      request.params.achievementId,
+      request.params.gameTypeId,
+      request.pre.getAuthUser.id
+    );
+
+    if (!achievement) {
+      throw Boom.notFound('achievement not found');
+    }
+
+    return h.response(achievement?.toJSON()).code(200);
+  } else {
+    const achievements = await findAllAchievementsByGameType(request.params.gameTypeId);
+    return h.response(arrayToJSON(achievements)).code(200);
+  }
+};
+
+export const upsertAchievementHandler: Lifecycle.Method = async (request, h) => {
+  const payload = request.payload as AchievementEditorData;
+
+  const game = await validateGameAuth(request.pre.getAuthUser.id, request.params.gameTypeId);
+
+  if (payload.id && game && !game._achievements?.map((a) => a.id).includes(payload.id)) {
+    throw Boom.forbidden(`achievement does not belong to gametypeId ${request.params.gameTypeId}`);
+  }
+
+  const [rslt] = await createOrUpdateAchievement({ ...payload }, request.params.gameTypeId);
+
+  return h.response(rslt?.toJSON()).code(200);
+};
+
+export const deleteAchievementHandler: Lifecycle.Method = async (request, h) => {
+  await validateGameAuth(request.pre.getAuthUser.id, request.params.gameTypeId);
+
+  const rslt = await deleteAchievementById(request.params.achievementId);
+
+  if (!rslt) {
+    throw Boom.notFound('achievement not found');
+  }
+
+  return h.response({ success: true }).code(200);
+};
