+++ src/modules/gameDevWebhook/webhookRoutes/leaderboardGameDevWebhookRoute.ts
@@ -1,26 +1,19 @@
 import type { ServerRoute } from '@hapi/hapi';
 
-import { appendUserToRequest } from '../../api-utils/appendUserToRequest';
-import {
-  postAchievementProgressResponseSchema,
-  postAchievementProgressRequestSchema,
-} from '../../api-utils/schemas/gameDev/achievementsSchemas';
+import { appendUserToRequest } from '../../../api-utils/appendUserToRequest';
 import {
   getLeaderboardRankResponseSchema,
   getUserLeaderboardResultScoreResponseSchema,
   postLeaderboardResultScoreResquestSchema,
   postLeaderboardResultScoreResponseSchema,
-} from '../../api-utils/schemas/gameDev/leaderboardSchemas';
-import { webhookValidation } from '../../api-utils/webhookValidations';
-
+} from '../../../api-utils/schemas/gameDev/leaderboardSchemas';
+import { webhookValidation } from '../../../api-utils/webhookValidations';
+import { parseWebhookPayload } from '../utils';
 import {
-  getAchievementsThruWebhookHandler,
-  postAchievementsProgressHandler,
   getUserLeaderboardResultHandler,
   postLeaderboardResultHandler,
   getLeaderboardRankHandler,
-} from './gameDevWebhookHandler';
-import { parseWebhookPayload } from './utils';
+} from '../webhookHandlers/leaderboardGameDevWebhookHandler';
 
 export const getGameLeaderboardResultRoute: ServerRoute = {
   method: 'GET',
@@ -94,58 +87,3 @@ export const postLeaderboardResultRoute: ServerRoute = {
   },
   handler: postLeaderboardResultHandler,
 };
-
-export const getGameAcheivmentsRoute: ServerRoute = {
-  method: 'GET',
-  path: '/webhooks/game-dev/achievements',
-  options: {
-    description: 'Get All the achievements related to a gameType',
-    tags: ['api'],
-    pre: [
-      {
-        method: webhookValidation,
-        assign: 'webhookValidation',
-      },
-    ],
-  },
-  handler: getAchievementsThruWebhookHandler,
-};
-
-export const postAchievementProgressRoute: ServerRoute = {
-  method: 'POST',
-  path: '/webhooks/game-dev/achievements/{achievementId}/progress',
-  options: {
-    description: 'Post user achievement progress',
-    payload: {
-      parse: false,
-      output: 'data',
-    },
-    tags: ['api'],
-    response: {
-      schema: postAchievementProgressResponseSchema,
-    },
-    pre: [
-      {
-        method: webhookValidation,
-        assign: 'webhookValidation',
-      },
-      {
-        method: parseWebhookPayload(postAchievementProgressRequestSchema),
-        assign: 'webhookPayload',
-      },
-      {
-        method: appendUserToRequest,
-        assign: 'appendUserToRequest',
-      },
-    ],
-  },
-  handler: postAchievementsProgressHandler,
-};
-
-export const gameDevWebhookRoutes: ServerRoute[] = [
-  getGameAcheivmentsRoute,
-  postAchievementProgressRoute,
-  getGameLeaderboardResultRoute,
-  getUserLeaderboardResultRoute,
-  postLeaderboardResultRoute,
-];
