--- src/modules/gameDevs/gameDevRoutes/index.ts
@@ -1,4 +1,0 @@
-export type { ServerRoute } from '@hapi/hapi';
-export { getAuthUser } from '../../../api-utils/getAuthUser';
-export { CAPABILITIES } from '../../../api-utils/interfaceAndTypes';
-export { gamedevGenericSchema } from '../../../api-utils/schemas/gameDev/game';
