+++ src/routes.ts
@@ -1,8 +1,8 @@
 import type { ServerRoute } from '@hapi/hapi';
 
-import { adminRoutes } from './modules/dashboard/adminRoutes';
-import { gameDevRoutes } from './modules/gameDevs/gameDevRoutes';
-import { gameDevWebhookRoutes } from './modules/gameDevs/gameDevWebhooksRoutes';
+import { adminRoutes } from './modules/dashboard/admin/adminRoutes';
+import { gameDevRoutes } from './modules/dashboard/gameDev/gameDevRoutes';
+import { gameDevWebhookRoutes } from './modules/gameDevWebhook/webhookRoutes';
 import { slackArenaRoutes } from './modules/slack/slackArenaRoute';
 import { slackRoutes } from './modules/slack/slackRoutes';
 import { slackTowerRoutes } from './modules/slack/slackTowerRoute';
