+++ test/modules/dashboard/gameDev/gameDevRoutes/leaderboardGameDevRoute.test.ts
@@ -1,0 +1,29 @@
+import { expect } from 'chai';
+import { RouteOptions } from '@hapi/hapi';
+import { getLeaderboardsRoute } from '../../../../../src/modules/dashboard/gameDev/gameDevRoutes/leaderboardGameDevRoutes';
+import { getLeaderboardHandler } from '../../../../../src/modules/dashboard/gameDev/gameDevHandlers/leaderboardGameDevHandler';
+import { multipleLeaderboardSchema } from '../../../../../src/api-utils/schemas/gameDev/leaderboardSchemas';
+import { CAPABILITIES } from '../../../../../src/api-utils/interfaceAndTypes';
+import { getAuthUser } from '../../../../../src/api-utils/getAuthUser';
+
+describe('getLeaderboardRoute', async () => {
+  it('should be configured as expected', async () => {
+    expect(getLeaderboardsRoute.method).to.equal('GET');
+    expect(getLeaderboardsRoute.path).to.equal(
+      '/dashboard/game-dev/games/{gameTypeId}/leaderboards'
+    );
+    expect(getLeaderboardsRoute.options?.bind).to.deep.equal({
+      requiredCapabilities: [CAPABILITIES.GAMEDEV_ACTIONS],
+    });
+    expect((getLeaderboardsRoute.options as RouteOptions).pre).to.deep.equal([
+      {
+        method: getAuthUser,
+        assign: 'getAuthUser',
+      },
+    ]);
+    expect((getLeaderboardsRoute.options as RouteOptions).response?.schema).to.equal(
+      multipleLeaderboardSchema
+    );
+    expect(getLeaderboardsRoute.handler).to.equal(getLeaderboardHandler);
+  });
+});
